Fix signedness error in glib/gspawn.c
glib/gspawn.c:2252:16: error: comparison of integer expressions of
  different signedness: ‘int’ and ‘gsize’ {aka ‘long unsigned int’}

 2252 |   if (argc + 2 > argv_buffer_len)
      |                ^

+++ glib/gspawn.c
@@ -2244,7 +2244,7 @@ script_execute (const gchar *file,
                 gchar      **envp)
 {
   /* Count the arguments.  */
-  int argc = 0;
+  gsize argc = 0;
   while (argv[argc])
     ++argc;
